Re-factoring
