+++ headline/mzamin.go
@@ -1,4 +1,4 @@
-package main
+package headline
 
 import (
 	"fmt"
