+++ utility.go
@@ -1,59 +1,30 @@
 package main
 
 import (
-	"bytes"
-	"io"
 	"log"
-	"net/http"
-	"strings"
 	"sync"
 	"time"
+
+	"github.com/shaharia-lab/headlines-go/headline"
 )
-
-type NewsClient interface {
-	GetHeadlines() (Response, error)
-	Name() string
-	SourceInfo() SourceInfo
-}
-
-type NewsItem struct {
-	Title string `json:"title"`
-	URL   string `json:"url"`
-}
-
-type SourceInfo struct {
-	Name     string `json:"name"`
-	Logo     string `json:"logo"`
-	Homepage string `json:"homepage"`
-}
-
-type Response struct {
-	Source    SourceInfo `json:"source"`
-	Headlines []NewsItem `json:"headlines"`
-}
-
-type CachedResponse struct {
-	Body      []Response
-	Timestamp time.Time
-}
 
 var (
 	headlinesCache sync.Map
 	cacheDuration  = 1 * time.Minute
 )
 
-func getHeadlines(sources []NewsClient) []Response {
+func getHeadlines(sources []headline.NewsClient) []headline.Response {
 	var wg sync.WaitGroup
-	results := make([]Response, len(sources))
+	results := make([]headline.Response, len(sources))
 
 	for i, source := range sources {
 		wg.Add(1)
-		go func(index int, s NewsClient) {
+		go func(index int, s headline.NewsClient) {
 			defer wg.Done()
 			items, err := s.GetHeadlines()
 			if err != nil {
 				log.Printf("Error fetching headlines from %s: %v", s.Name(), err)
-				results[index] = Response{Source: s.SourceInfo(), Headlines: nil}
+				results[index] = headline.Response{Source: s.SourceInfo(), Headlines: nil}
 			} else {
 				results[index] = items
 			}
@@ -64,9 +35,9 @@ func getHeadlines(sources []NewsClient) []Response {
 	return results
 }
 
-func getCachedHeadlines() ([]Response, bool) {
+func getCachedHeadlines() ([]headline.Response, bool) {
 	if cachedResp, ok := headlinesCache.Load("headlines"); ok {
-		cached := cachedResp.(CachedResponse)
+		cached := cachedResp.(headline.CachedResponse)
 		if time.Since(cached.Timestamp) < cacheDuration {
 			return cached.Body, true
 		}
@@ -74,72 +45,9 @@ func getCachedHeadlines() ([]Response, bool) {
 	return nil, false
 }
 
-func cacheHeadlines(headlines []Response) {
-	headlinesCache.Store("headlines", CachedResponse{
+func cacheHeadlines(headlines []headline.Response) {
+	headlinesCache.Store("headlines", headline.CachedResponse{
 		Body:      headlines,
 		Timestamp: time.Now(),
 	})
-}
-
-type CachingHTTPClient struct {
-	client    *http.Client
-	cache     sync.Map
-	userAgent string
-}
-
-func NewCachingHTTPClient(timeout time.Duration, userAgent string) *CachingHTTPClient {
-	return &CachingHTTPClient{
-		client: &http.Client{
-			Timeout: timeout,
-		},
-		userAgent: userAgent,
-	}
-}
-
-func (c *CachingHTTPClient) Get(url string) (*http.Response, error) {
-	if cachedBody, ok := c.cache.Load(url); ok {
-		return &http.Response{
-			StatusCode: http.StatusOK,
-			Body:       io.NopCloser(strings.NewReader(cachedBody.(string))),
-		}, nil
-	}
-
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Set("User-Agent", c.userAgent)
-
-	resp, err := c.client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		resp.Body.Close()
-		return nil, err
-	}
-	resp.Body.Close()
-
-	c.cache.Store(url, string(body))
-
-	return &http.Response{
-		StatusCode: resp.StatusCode,
-		Body:       io.NopCloser(bytes.NewReader(body)),
-	}, nil
-}
-
-func completeURL(baseURL, relativeURL string) string {
-	if relativeURL == "" {
-		return ""
-	}
-	if strings.HasPrefix(relativeURL, "http") {
-		return relativeURL
-	}
-	baseURL = strings.TrimRight(baseURL, "/")
-	if strings.HasPrefix(relativeURL, "/") {
-		return baseURL + relativeURL
-	}
-	return baseURL + "/" + relativeURL
 }
